+++ example/states/todoList/useTodoList.ts
@@ -1,4 +1,4 @@
-import { createOne } from '../../../src';
+import { create } from '../../../src';
 import { TodoItemType } from './types';
 
 let initialState: Readonly<TodoItemType>[] = [];
@@ -7,7 +7,7 @@ export type TodoListStateType = Readonly<typeof initialState>;
 
 export const todoListInitialState: TodoListStateType = initialState;
 
-const [useTodoList, todoListStore] = createOne<TodoListStateType>(
+const [useTodoList, todoListStore] = create<TodoListStateType>(
   todoListInitialState
 );
 export { useTodoList, todoListStore };