+++ example/states/useCount.ts
@@ -1,6 +1,6 @@
 /** Generator by `use-one-templates` https://github.com/suhaotian/use-one-templates */
 
-import { createOne } from '../../src';
+import { create } from '../../src';
 
 let initialState = 0;
 
@@ -8,7 +8,7 @@ export type CountStateType = Readonly<typeof initialState>;
 
 export const countInitialState: CountStateType = initialState;
 
-const [useCount, countStore] = createOne<CountStateType>(countInitialState);
+const [useCount, countStore] = create<CountStateType>(countInitialState);
 export { useCount, countStore };
 
 export const countSelectors = {