+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from './createOne';
+export * from './create';