+++ src/types.ts
@@ -1,7 +1,7 @@
 export type ReadonlyNonBasic<T> = T extends object ? Readonly<T> : T;
 
 /** the create one options */
-export type CreateOneOptions = {
+export type createOptions = {
   /** useEffect or useLayoutEffect */
   useEffect: boolean;
   /** state name */